Initial UI Layer
Orthographic overlay for a simple panel.  Still need resize and a few widdes as well as mouse capture/tracking etc to go with the sections.
Rearchitected how the parts go together for a window/view/scene/etc/
